6340 | 64 files | M DumpSource2/schemas/client/CDOTA_Ability_Dawnbreaker_Celestial_Hammer.h,  M DumpSource2/schemas/client/CDOTA_Ability_Hoodwink_Decoy.h,  M DumpSource2/schemas/client/CDOTA_Modifier_Abaddon_BorrowedTime.h,  M DumpSource2/schemas/client/CDOTA_Modifier_Beastmaster_DrumsOfSlom.h,  M DumpSource2/schemas/client/CDOTA_Modifier_DarkSeer_WallOfReplica.h,  M DumpSource2/schemas/client/CDOTA_Modifier_DarkWillow_Bedlam.h,  M DumpSource2/schemas/client/CDOTA_Modifier_DarkWillow_Terrorize_Thinker.h,  M DumpSource2/schemas/client/CDOTA_Modifier_Dawnbreaker_Fire_Wreath_Caster.h,  M DumpSource2/schemas/client/CDOTA_Modifier_Dazzle_Innate_Weave_Armor.h,  M DumpSource2/schemas/client/CDOTA_Modifier_Invoker_ColdSnap.h,  M DumpSource2/schemas/client/CDOTA_Modifier_Item_UnrelentingEye.h,  M DumpSource2/schemas/client/CDOTA_Modifier_KeeperOfTheLight_Will_O_Wisp_Aura.h,  M DumpSource2/schemas/client/CDOTA_Modifier_Leshrac_Lightning_Storm_Slow.h,  M DumpSource2/schemas/client/CDOTA_Modifier_Luna_MoonGlaive.h,  M DumpSource2/schemas/
https://steamdb.info/patchnotes/17593677/

+++ DumpSource2/schemas/client/CDOTA_Modifier_Luna_MoonGlaive.h
@@ -1,6 +1,6 @@
 class CDOTA_Modifier_Luna_MoonGlaive
 {
-	int32 range;
+	float32 range;
 	int32 bounces;
 	int32 damage_reduction_percent;
 	ParticleIndex_t m_nMoonGlaiveFXIndex;
