+++ DumpSource2/schemas/client/CDOTA_Modifier_Sven_GreatCleave.h
@@ -1,9 +1,9 @@
 class CDOTA_Modifier_Sven_GreatCleave
 {
-	int32 cleave_starting_width;
-	int32 cleave_ending_width;
-	int32 cleave_distance;
-	int32 great_cleave_damage;
+	float32 cleave_starting_width;
+	float32 cleave_ending_width;
+	float32 cleave_distance;
+	float32 great_cleave_damage;
 	int32 strength_bonus;
 	int16 m_nLastCleaveRecord;
 	int32 m_nLastCleaveKills;
