+++ DumpSource2/schemas/client/CDOTA_Modifier_Terrorblade_Metamorphosis_Fear_Thinker.h
@@ -5,8 +5,8 @@ class CDOTA_Modifier_Terrorblade_Metamorphosis_Fear_Thinker
 	GameTime_t m_fLastThink;
 	ParticleIndex_t m_nFXIndex;
 	CUtlVector< CHandle< C_BaseEntity > > m_EntitiesHit;
-	int32 scepter_speed;
-	int32 scepter_radius;
+	float32 scepter_speed;
+	float32 scepter_radius;
 	float32 damage;
 	bool m_bFirstThink;
 };
