+++ DumpSource2/schemas/client/CDOTA_Modifier_Visage_SummonFamiliars_StoneForm_Thinker.h
@@ -1,7 +1,7 @@
 class CDOTA_Modifier_Visage_SummonFamiliars_StoneForm_Thinker
 {
-	int32 stun_radius;
-	int32 stun_damage;
+	float32 stun_radius;
+	float32 stun_damage;
 	float32 stun_delay;
 	float32 stun_duration;
 	float32 stone_duration;
