+++ DumpSource2/schemas/client/C_DOTA_Ability_Elder_Titan_EchoStomp.h
@@ -4,8 +4,8 @@ class C_DOTA_Ability_Elder_Titan_EchoStomp
 	ParticleIndex_t m_nFXIndexSpirit;
 	ParticleIndex_t m_nFXIndexTitanB;
 	ParticleIndex_t m_nFXIndexSpiritB;
-	int32 radius;
-	int32 stomp_damage;
+	float32 radius;
+	float32 stomp_damage;
 	float32 sleep_duration;
 	float32 cast_time;
 	CUtlVector< CHandle< C_BaseEntity > > m_vecStompedHeroes;
