+++ DumpSource2/schemas/server/CDOTA_Ability_Hoodwink_Decoy.h
@@ -2,7 +2,7 @@ class CDOTA_Ability_Hoodwink_Decoy
 {
 	CUtlVector< CHandle< CBaseEntity > > hIllusion;
 	float32 decoy_stun_duration;
-	int32 decoy_detonate_radius;
-	int32 projectile_speed;
+	float32 decoy_detonate_radius;
+	float32 projectile_speed;
 	ParticleIndex_t m_nFXIndex;
 };
