+++ DumpSource2/schemas/server/CDOTA_Modifier_DarkSeer_WallOfReplica.h
@@ -3,7 +3,7 @@ class CDOTA_Modifier_DarkSeer_WallOfReplica
 	Vector m_vWallDirection;
 	Vector m_vWallRight;
 	CUtlOrderedMap< int32, GameTime_t > m_PreventReplicateTime;
-	int32 width;
+	float32 width;
 	int32 wall_damage;
 	float32 slow_duration;
 	int32 replica_damage_incoming;
