+++ DumpSource2/schemas/server/CDOTA_Modifier_DarkWillow_Terrorize_Thinker.h
@@ -1,9 +1,9 @@
 class CDOTA_Modifier_DarkWillow_Terrorize_Thinker
 {
 	CHandle< CBaseEntity > m_hWisp;
-	int32 destination_travel_speed;
-	int32 return_travel_speed;
-	int32 destination_radius;
+	float32 destination_travel_speed;
+	float32 return_travel_speed;
+	float32 destination_radius;
 	float32 impact_damage;
 	float32 destination_status_duration;
 	float32 initial_delay;
