+++ DumpSource2/schemas/server/CDOTA_Modifier_Silencer_BrainDrain.h
@@ -1,5 +1,5 @@
 class CDOTA_Modifier_Silencer_BrainDrain
 {
-	int32 permanent_int_steal_range;
+	float32 permanent_int_steal_range;
 	int32 permanent_int_steal_amount;
 };
