+++ DumpSource2/schemas/server/CDOTA_Modifier_Tiny_Tree_Channel.h
@@ -1,8 +1,8 @@
 class CDOTA_Modifier_Tiny_Tree_Channel
 {
-	int32 tree_grab_radius;
-	int32 splash_radius;
-	int32 speed;
+	float32 tree_grab_radius;
+	float32 splash_radius;
+	float32 speed;
 	float32 interval;
 	Vector m_vPos;
 };
