+++ DumpSource2/schemas/server/CDOTA_Modifier_VoidSpirit_AetherRemnantThinker.h
@@ -1,6 +1,6 @@
 class CDOTA_Modifier_VoidSpirit_AetherRemnantThinker
 {
-	int32 radius;
+	float32 radius;
 	Vector m_vFacingDir;
 	CHandle< CBaseEntity > m_hRemnant;
 	bool m_bTriggered;
